rename seg002_45ea() and seg002_4658()
I managed to find the purpose of these two functions.
With them you select an item in a shop view.
If a mouse is enabled, you can only select items with the mouse.
If no mouse is enabled, you have to use the keyboard.

So I had to disable the mouse to test the already rewritten keyboard function.
And found a bug introduced by me.
This is fixed now.
Also host_*() functions are used in the mouse function.