Fix VC++2008 Debug version
Debug-Builds with VC++2008 produce a warning message when the
NVF loader process_nvf() is executed with a RLE image.

The reason is that, in this case, the return value was neither set nor needed.
Since Debug-Builds can be useful, we will not stress ourselfs with
clicking every warning message away.