add host file functions
Here we have two functions, which operate on files and
return a pointer to struct FILE, instead of a DOS-handle.

These functions should be used in rewritten code now,
that we may get rid of the Borland C Code.

fd_open_datfile() works a bit different, but it works
compared to the original. The original did a loop if
DSAGEN.DAT could not be opened and printed a message in
an infobox. Sadly the background is in DSAGEN.DAT,
so the program just crashes.