M302de: Original-Bugfix: check before enable Umbrik
This is a _should read before write_ Bug.
A 0 in this variable means that this informant is not known.
A 1 in this variable means that this informant is known, but not served.
A 2 in this variable means that this informant is known and served.

Writing blindly 1 into it would erase a 2, which may stand in it.